mos-components-ci/componentspython/requirements.txt
python-glanceclient>=0.17.1
python-keystoneclient>=1.3.0
python-novaclient>=2.22.0
joblib
launchpadlib
requests>=2.5.2
jenkinsapi>=0.2.25
python-subunit>=0.0.18
testrepository>=0.0.18
PyYAML>=3.1.0
libvirt-python>=1.2.5
pbr>=0.5.21,<1.0
